backend/requirements.txt
appdirs==1.4.4
APScheduler==3.8.1
asgiref==3.3.1
attrs==21.2.0
Authlib==0.15.4
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
click==7.1.2
cryptography==3.4.8
decorator==5.1.0
defusedxml==0.7.1
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-model-utils==4.1.1
django-notifications-hq==1.6.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
dnspython==2.1.0
docutils==0.17.1
et-xmlfile==1.1.0
exchangelib==3.3.2
Flask==2.0.1
fpdf2==2.4.3
furl==2.1.2
graphviz==0.13.2
idna==2.10
importlib-metadata==3.7.0
isodate==0.6.0
itsdangerous==2.0.1
java-access-bridge-wrapper==0.7.5
Jinja2==3.0.1
joblib==1.0.1
jsonfield==3.1.0
jsonpath-ng==1.5.3
jsonschema==3.2.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
mss==6.1.0
mysqlclient==2.0.3
netsuitesdk==1.24.0
nltk==3.5
notifiers==1.2.1
ntlm-auth==1.5.0
numpy==1.21.2
oauthlib==3.1.1
openpyxl==3.0.9
orderedmultidict==1.0.1
pandas==1.1.5
pdfminer.six==20201018
Pillow==8.1.2
platformdirs==2.3.0
ply==3.11
pycparser==2.20
Pygments==2.10.0
PyJWT==1.7.1
pynput-robocorp-fork==4.0.0
pyobjc-core==7.3
pyobjc-framework-Cocoa==7.3
pyobjc-framework-Quartz==7.3
PyPDF2==1.26.0
pyperclip==1.8.2
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.15.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
regex==2021.3.17
requests==2.25.1
requests-file==1.5.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3987==1.3.8
robocorp-dialog==0.4.1
robotframework==4.1.1
robotframework-pythonlibcore==3.0.0
robotframework-requests==0.9.2
robotframework-seleniumlibrary==5.1.3
robotframework-seleniumtestability==1.1.0
rpaframework==11.2.1
rpaframework-core==6.4.0
rpaframework-dialogs==0.4.2
rpaframework-pdf==0.8.0
selenium==3.141.0
simple-salesforce==1.11.4
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.1
swapper==1.1.2.post1
tqdm==4.59.0
tweepy==3.10.0
typing-extensions==3.7.4.3
tzdata==2021.1
tzlocal==4.1
urllib3==1.26.5
webdrivermanager==0.10.0
Werkzeug==2.0.1
wrapt==1.12.1
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
zeep==4.1.0
zipp==3.4.0
